refactor: rename /guards/ to /middleware/
